manpages: fixup synopsis
live-build page: list of individual commands outdated (color options not
added); incorrect (conffile is `lb config` specific as to be addressed
otherwise separately); is unnecessary here; and the synopsis suggests
usage without a command (e.g. `lb --help` is wrong when it is not).

lb: consistensy fix in formatting.

all: placeholders should be in capitals for consistency; "OPTIONS" is
better the "live-build options" as a placeholder.

config: unlike the others where "live-build options" is replaced with
"OPTIONS", it makes more sense to specifically list the info (help; usage;
version} options, since below the full list of other options including
other generic options are given.

Gbp-Dch: Short