dracut: Support diverted configuration files of Calamares
When a configuration file for Calamares is diverted by e.g.
`config-package-dev`, it becomes a symlink. The in-place modifications
must be applied on the symlinked file, not on a newly created copy.
After removal of the package, no remainders are left.

See https://salsa.debian.org/live-team/live-build/-/merge_requests/353#note_540394