QEMU Support: Fix garbled output during text boot
When the live system is booted in QEMU with the 'VGA' driver, the
console output is garbled until the 'bochs' kernel module is loaded.
When the 'verify-checksums' kernel command line option is used, the
systemd 'modprobe@drm.service' has not been run yet, which would use the
kernel module from the live system. Therefore the kernel module is
placed inside the initramfs.